chore: rename component dir
